+++ helpers/generate_question_helpers.py
@@ -1,10 +1,19 @@
 from sqlalchemy.orm import Session
 
-from database.db_models import Quiz, Question
+from models.db_models import Quiz, Question
 from starlette.exceptions import HTTPException
 from helpers import vectorstore_helpers
 import uuid
 from cloudwatch_logger import cloudwatch_logger
+
+from models.llm import GeneratedQuestions
+
+from langchain.output_parsers.openai_tools import JsonOutputToolsParser
+from langchain_core.runnables import RunnablePassthrough
+from langchain_core.prompts import ChatPromptTemplate
+
+from langchain_openai import ChatOpenAI
+
 
 CHAIN_CACHE = {}
 
@@ -30,24 +39,6 @@ def get_generated_questions(user_id: str, question_generation_settings, db: Sess
         q['question_id'] = str(uuid.uuid4())
     return questions
  
-
-
-def _parse_generation_settings(question_generation_settings, user_id: str, db: Session):
-    try:
-        quiz_id = question_generation_settings.get('quiz_id')
-        amount = question_generation_settings.get('amount')
-        round_specific_instructions = question_generation_settings.get('instructions')
-        keywords = question_generation_settings.get('keywords')
-
-        quiz = db.query(Quiz).filter(Quiz.user_id == user_id, Quiz.quiz_id == quiz_id).first()
-
-        cloudwatch_logger.info(f' Succesfully parsed question generation settings')
-        return quiz, amount, round_specific_instructions, keywords
-    except Exception as e:
-        cloudwatch_logger.error(f'Error while parsing question generating settings.'
-                                f'Details: {str(e)}')
-        raise e
-
 
 def _get_prompt(amount, quiz: Quiz, existing_questions: list[Question], keywords=None,
                 round_specific_instructions=None):
@@ -82,6 +73,59 @@ def _get_prompt(amount, quiz: Quiz, existing_questions: list[Question], keywords
     return prompt
 
 
+def _get_chain(quiz: Quiz | None):
+    global CHAIN_CACHE
+    if quiz.quiz_id in CHAIN_CACHE:
+        # Get chain from cache
+        chain = CHAIN_CACHE[quiz.quiz_id]
+    else:
+        # Initialize chain and cache
+        chain = get_conversation_chain(quiz=quiz)
+        CHAIN_CACHE[quiz.quiz_id] = chain
+    return chain
+
+
+def get_conversation_chain(quiz: Quiz):
+    cloudwatch_logger.info(f'Attempting to get conversation chain for the quiz ID: {quiz.quiz_id}')
+    retriever = vectorstore_helpers.get_retriever(quiz=quiz)
+
+    template = """Answer the question based only on the following context:
+    {context}
+
+    Question: {question}
+    """
+    prompt = ChatPromptTemplate.from_template(template)
+    model = ChatOpenAI().bind_tools([GeneratedQuestions])
+
+    chain = (
+        {"context": retriever, 'question': RunnablePassthrough()}
+        | prompt
+        | model
+        | JsonOutputToolsParser()
+    )
+
+    cloudwatch_logger.info(f'Chain successfully created')
+
+    return chain
+
+
+def _parse_generation_settings(question_generation_settings, user_id: str, db: Session):
+    try:
+        quiz_id = question_generation_settings.get('quiz_id')
+        amount = question_generation_settings.get('amount')
+        round_specific_instructions = question_generation_settings.get('instructions')
+        keywords = question_generation_settings.get('keywords')
+
+        quiz = db.query(Quiz).filter(Quiz.user_id == user_id, Quiz.quiz_id == quiz_id).first()
+
+        cloudwatch_logger.info(f' Succesfully parsed question generation settings')
+        return quiz, amount, round_specific_instructions, keywords
+    except Exception as e:
+        cloudwatch_logger.error(f'Error while parsing question generating settings.'
+                                f'Details: {str(e)}')
+        raise e
+
+
 def _validate_inputs_and_authorization(quiz: Quiz | None, amount: str):
     if amount is None:
         raise HTTPException(status_code=400, detail="Bad request.")
@@ -91,17 +135,6 @@ def _validate_inputs_and_authorization(quiz: Quiz | None, amount: str):
     return True
 
 
-def _get_chain(quiz: Quiz | None):
-    global CHAIN_CACHE
-    if quiz.quiz_id in CHAIN_CACHE:
-        # Get chain from cache
-        chain = CHAIN_CACHE[quiz.quiz_id]
-    else:
-        # Initialize chain and cache
-        chain = vectorstore_helpers.get_conversation_chain(quiz=quiz)
-        CHAIN_CACHE[quiz.quiz_id] = chain
-    return chain
-
-
 def _parse_questions(response):
     return response[0].get('args').get('questions')
+
