+++ helpers/vectorstore_helpers.py
@@ -1,15 +1,9 @@
 import os
-from typing import List
 from fastapi import UploadFile
-from database.db_models import Quiz
+from models.db_models import Quiz
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_openai import OpenAIEmbeddings
 from langchain_pinecone import Pinecone
-from langchain.output_parsers.openai_tools import JsonOutputToolsParser
-from langchain_core.runnables import RunnablePassthrough
-from langchain_core.prompts import ChatPromptTemplate
-from langchain_core.pydantic_v1 import BaseModel, Field
-from langchain_openai import ChatOpenAI
 from loaders import PDFMinerPagesLoader
 from cloudwatch_logger import cloudwatch_logger
 import time
@@ -24,6 +18,9 @@ def add_quiz_to_vectorstore(source_file: UploadFile, new_quiz: Quiz, file_hash: 
     for attempt in range(MAX_RETRY_ATTEMPTS):
         try:
             documents = get_documents_from_file(source_file=source_file, file_hash=file_hash)
+            print('=======')
+            print(documents)
+            print('=======')
             cloudwatch_logger.info(f'Created documents from the file.')
             index_name, namespace, embeddings = get_pinecone_config(quiz=new_quiz)
             cloudwatch_logger.info(f'Set up pinecone config.')
@@ -66,7 +63,7 @@ def get_documents_from_file(source_file: UploadFile, file_hash: str):
     return split_documents
 
 
-def _get_retriever(quiz: Quiz):
+def get_retriever(quiz: Quiz):
     try:
         index_name, namespace, embeddings = get_pinecone_config(quiz=quiz)
         cloudwatch_logger.info(f'Getting retriever index name: {index_name}')
@@ -82,40 +79,3 @@ def _get_retriever(quiz: Quiz):
         raise e
 
 
-class GeneratedQuestion(BaseModel):
-    question_id: str = Field(description='A unique identifier for each generated question.')
-    question_text: str = Field(description="The question to be added to the quiz")
-    multiple_choices: str = Field(description="A string that includes four, comma separated potential answers to the "
-                                              "question")
-    correct_answer: str = Field(description='Correct answer to the question. Must be one of the four multiple choices.')
-    difficulty: str = Field(description='The proposed difficulty of the question. Only possible options are easy, medium, hard')
-    score: str = Field(description='The score of the question, defining the question quality. A value from 0 to 5, 5 being the best quality.')
-    question_type: str = Field(description='Always set it as "multi"')
-    
-
-class GeneratedQuestions(BaseModel):
-    questions: List[GeneratedQuestion] = Field(description="List of generated questions in the specified amount")
-
-
-def get_conversation_chain(quiz: Quiz):
-    cloudwatch_logger.info(f'Attempting to get conversation chain for the quiz ID: {quiz.quiz_id}')
-    retriever = _get_retriever(quiz=quiz)
-
-    template = """Answer the question based only on the following context:
-    {context}
-
-    Question: {question}
-    """
-    prompt = ChatPromptTemplate.from_template(template)
-    model = ChatOpenAI().bind_tools([GeneratedQuestions])
-
-    chain = (
-        {"context": retriever, 'question': RunnablePassthrough()}
-        | prompt
-        | model
-        | JsonOutputToolsParser()
-    )
-
-    cloudwatch_logger.info(f'Chain successfully created')
-
-    return chain
