+++ middleware.py
@@ -9,7 +9,7 @@ import json
 def setup_cors_middleware(app) -> None:
     app.add_middleware(
         CORSMiddleware,
-        # allow_origins=['*'],
+        allow_origins=['*'],
         allow_credentials=True,
         allow_headers=['Authorization'],
         allow_methods=['*']
